starting image.js; switch to circle
