+++ .gitignore
@@ -1,8 +1,7 @@
 .DS_Store
+
+/.idea/
 /.sass-cache/
-/bower_components/
-/node_modules/
 /build/
-.idea/
 /demos/local
-/ac
+/node_modules/
