+++ Makefile
@@ -2,6 +2,6 @@ install:
 	obt install --verbose
 
 verify:
-	nbt verify --skip-layout-checks --skip-dotenv-check
+	obt verify
 
 test: verify
