+++ helpers/index.js
@@ -1,4 +1,4 @@
 import buildUrl from '../helpers/build-url';
 import addWidth from '../helpers/add-width';
 
-export {buildUrl, addWidth};
+export { buildUrl, addWidth };
