+++ main.js
@@ -1,0 +1,4 @@
+import Image from './templates/image';
+import Picture from './templates/picture';
+
+export { Image, Picture };
